+++ backend/src/main/kotlin/com/finance/backend/remit/request/RemitPhoneReq.kt
@@ -15,7 +15,7 @@ class RemitPhoneReq {
     @JsonProperty("ac_receive")
     val acReceive: String = ""
 
-    val value: Int = 0
+    val value: Long = 0
 
     val receive: String = ""
